Donw with comments

+++ TodoListProject/src/main/java/com/todolist/servicesimpl/TaskServiceImpl.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.todolist.dao.TaskDao;
+import com.todolist.exceptions.ServicesException;
 import com.todolist.models.Task;
 import com.todolist.services.TaskService;
 
@@ -17,32 +18,69 @@ public class TaskServiceImpl implements TaskService {
 
 	@Override
 	public Task addTask(Task task) {
-		return taskDao.save(task);
+		Task taskDetails = null;
+		try {
+			taskDetails = taskDao.save(task);
+		} catch (Exception e) {
+			throw new ServicesException("Exception occurred while saving task details to table.");
+		}
+		return taskDetails;
+
 	}
 
 	@Override
 	public Task getTaskById(Integer id) {
-		return taskDao.getById(id);
+		Task taskDetails = null;
+		try {
+			taskDetails = taskDao.getById(id);
+		} catch (Exception e) {
+			throw new ServicesException("Task Not Found : " + id);
+
+		}
+		return taskDetails;
 	}
 
 	@Override
 	public Task editTask(Task task) {
-		return taskDao.save(task);
+		Task taskDetails = null;
+		try {
+			taskDetails = taskDao.save(task);
+		} catch (Exception e) {
+			throw new ServicesException("Exception occurred while editing task details.");
+		}
+		return taskDetails;
 	}
 
 	@Override
 	public void deleteTask(Integer id) {
-		taskDao.deleteById(id);
+		if (id == null || id < 0) {
+			throw new ServicesException("Exception occured due to invalid task id.");
+		} else {
+			taskDao.deleteById(id);
+		}
+
 	}
 
 	@Override
 	public List<Task> findByUsername(String username) {
-		return taskDao.findByUsername(username);
+		List<Task> tasks = null;
+		try {
+			tasks = taskDao.findByUsername(username);
+		} catch (Exception e) {
+			throw new ServicesException("Exception occurred while Fetching tasks details.");
+		}
+		return tasks;
 	}
 
 	@Override
 	public List<Task> getTasks() {
-		return taskDao.findAll();
+		List<Task> tasks = null;
+		try {
+			tasks = taskDao.findAll();
+		} catch (Exception e) {
+			throw new ServicesException("Exception occurred while Fetching tasks details.");
+		}
+		return tasks;
 	}
 
 }
